refactor(viewer): hoist canonical quantile label helper (R4)
Three places canonicalized percentile/quantile labels to "pXX" form —
compare.js (scatter split label), viewer_core.js (baseline scatter
normalizer), and viewer_core.js (experiment scatter extractor). The
three implementations had subtle differences and would drift. Hoist
to canonicalQuantileLabel in util/compare_math.js and use it from all
three sites.
